get rid of the f16 mess

+++ assets/shaders/instancing.wgsl
@@ -1,17 +1,17 @@
 #import bevy_pbr::mesh_functions::{get_world_from_local, mesh_position_local_to_clip}
 
 struct Vertex {
-    @location(0) position: vec3<f16>,
-    @location(1) normal: vec3<f16>,
-    @location(2) uv: vec2<f16>,
+    @location(0) position: vec3<f32>,
+    @location(1) normal: vec3<f32>,
+    @location(2) uv: vec2<f32>,
 
-    @location(3) i_pos_scale: vec4<f16>,
-    @location(4) i_color: vec4<f16>,
+    @location(3) i_pos_scale: vec4<f32>,
+    @location(4) i_color: vec4<f32>,
 };
 
 struct VertexOutput {
-    @builtin(position) clip_position: vec4<f16>,
-    @location(0) color: vec4<f16>,
+    @builtin(position) clip_position: vec4<f32>,
+    @location(0) color: vec4<f32>,
 };
 
 @vertex
@@ -31,6 +31,6 @@ fn vertex(vertex: Vertex) -> VertexOutput {
 }
 
 @fragment
-fn fragment(in: VertexOutput) -> @location(0) vec4<i8> {
+fn fragment(in: VertexOutput) -> @location(0) vec4<f32> {
     return in.color;
 }
